Move LockFile and related types to a separate file
